requirements.txt
setuptools
python-wappalyzer
google-generativeai
virustotal_python
colorama
flask_cors
flask
lxml
fuzzywuzzy
prettytable
